Simplify core API
