+++ .changeset/polite-buses-call.md
@@ -1,0 +1,7 @@
+---
+'@svg-use/webpack': minor
+'@svg-use/rollup': minor
+'@svg-use/core': minor
+---
+
+Refactor @svg-use/core API, update plugins to use it.
