+++ packages/core/src/getSvgIdAttribute.ts
@@ -1,0 +1,8 @@
+export type GetSvgIdFunction = ({
+  existingId,
+}: {
+  existingId?: string;
+}) => string;
+
+export const defaultGetSvgIdAttribute: GetSvgIdFunction = () =>
+  'use-href-target';
