--- packages/core/src/idCreationFunction.ts
@@ -1,7 +1,0 @@
-export type IdCreationFunction = ({
-  filename,
-  existingId,
-}: {
-  filename: string;
-  existingId?: string;
-}) => string;
