+++ packages/core/src/theme/svgoMakeThemeable.ts
@@ -1,0 +1,23 @@
+import type { CustomPlugin } from 'svgo';
+import {
+  GetThemeSubstitutionFunction,
+  xastMakeThemeable,
+} from './makeThemeable.js';
+
+/**
+ * Substitute hardcoded color values with other ones (usually custom
+ * properties). SVGO-compatible version of {@link xastMakeThemeable}
+ */
+export const svgoMakeThemeable = (
+  themeSubstitutionFunction: GetThemeSubstitutionFunction,
+): CustomPlugin => ({
+  name: 'make-themeable',
+  fn: (root) => {
+    // XastRoot does not line up with Root in the types, but they are both xast
+    // roots, so the cast is safe. makeThemeable does its own visiting of the
+    // tree (in two passes), so we do not use SVGO's visitor.
+    // @ts-expect-error -- the types do not line up, as above
+    xastMakeThemeable(root, themeSubstitutionFunction);
+    return null;
+  },
+});
